Recompose ShazamResultView into TrackView and TrackActionsView
